+++ api/jobs_test.go
@@ -243,6 +243,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -287,6 +288,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(30 * time.Second),
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(30 * time.Second),
@@ -334,6 +338,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -367,6 +372,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(5 * time.Second),
 							MaxDelay:      timeToPtr(0),
 							Unlimited:     boolToPtr(false),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Tasks: []*Task{
 							{
@@ -408,6 +416,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -452,6 +461,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(30 * time.Second),
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(30 * time.Second),
@@ -575,6 +587,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Type:              stringToPtr("service"),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -621,7 +634,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Migrate: boolToPtr(false),
 							SizeMB:  intToPtr(300),
 						},
-
+						Consul: &Consul{
+							Namespace: "",
+						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(30 * time.Second),
 							MaxParallel:      intToPtr(1),
@@ -736,6 +751,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -788,6 +804,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 				TaskGroups: []*TaskGroup{
 					{
 						Name: stringToPtr("bar"),
+						Consul: &Consul{
+							Namespace: "",
+						},
 						Update: &UpdateStrategy{
 							Stagger:        timeToPtr(2 * time.Second),
 							MaxParallel:    intToPtr(2),
@@ -823,6 +842,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -868,6 +888,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
 						},
+						Consul: &Consul{
+							Namespace: "",
+						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(2 * time.Second),
 							MaxParallel:      intToPtr(2),
@@ -911,6 +934,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(30 * time.Second),
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(1 * time.Second),
@@ -971,6 +997,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Interval: timeToPtr(30 * time.Minute),
 							Mode:     stringToPtr("fail"),
 						},
+						Consul: &Consul{
+							Namespace: "",
+						},
 						Tasks: []*Task{
 							{
 								Name: "task1",
@@ -989,6 +1018,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),
@@ -1033,6 +1063,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(30 * time.Second),
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(30 * time.Second),
@@ -1082,6 +1115,9 @@ func TestJobs_Canonicalize(t *testing.T) {
 							Delay:         timeToPtr(30 * time.Second),
 							MaxDelay:      timeToPtr(1 * time.Hour),
 							Unlimited:     boolToPtr(true),
+						},
+						Consul: &Consul{
+							Namespace: "",
 						},
 						Update: &UpdateStrategy{
 							Stagger:          timeToPtr(30 * time.Second),
@@ -1153,6 +1189,7 @@ func TestJobs_Canonicalize(t *testing.T) {
 				Priority:          intToPtr(50),
 				AllAtOnce:         boolToPtr(false),
 				ConsulToken:       stringToPtr(""),
+				ConsulNamespace:   stringToPtr(""),
 				VaultToken:        stringToPtr(""),
 				VaultNamespace:    stringToPtr(""),
 				NomadTokenID:      stringToPtr(""),